Refactor availability management and fix time update
