+++ src/app/features/teachers/utils/time.utils.ts
@@ -1,0 +1,22 @@
+export function formatTimeDisplay(time: string): string {
+  const [hourStr, minuteStr] = time.split(':');
+  const hour = parseInt(hourStr, 10);
+  const period = hour >= 12 ? 'PM' : 'AM';
+  const displayHour = hour > 12 ? hour - 12 : (hour === 0 ? 12 : hour);
+  return `${displayHour}:${minuteStr} ${period}`;
+}
+
+export function calculateDuration(startTime: string, endTime: string): string {
+  const [startHour, startMinute] = startTime.split(':').map(Number);
+  const [endHour, endMinute] = endTime.split(':').map(Number);
+  const startMinutes = startHour * 60 + startMinute;
+  const endMinutes = endHour * 60 + endMinute;
+  const durationMinutes = endMinutes - startMinutes;
+  const hours = Math.floor(durationMinutes / 60);
+  const minutes = durationMinutes % 60;
+  return hours > 0 ? `${hours}h ${minutes > 0 ? minutes + 'min' : ''}` : `${minutes}min`;
+}
+
+export function toSeconds(time: string): string {
+  return time.length === 5 ? `${time}:00` : time;
+}
